requirements.txt
google-api-python-client
pandas
python-gnupg
xmltodict
